Added big-endian platform build error to CoreKit.
The new integer abstraction is nice and simple because it is allowd to reinterpret memoy. As a result, I need each integer to be correctly laid out in memory. While it is possible to support big endianness, it requires some work that I have not gotten around to yet. Having said all of that, we all know that little endian is the true endian and that banning big endian platform does everybody a favor.