Replaced: DataInt/drop(as:) with DataInt/next(as:).
It is true that DataInt/drop(as:) is more flexible, but it is also less purposful. Note that the subscript is even more powerful than drop(as:), so it's only a small convenience. I think it's better to offer an iterator-like api, because that matchs the purpose of most drop(as:) use cases.