Reworked: normalization w.r.t. TextInt and InfiniInt.
I noticed that it would be better to normalize elements in InfiniInt.init(load:) than in TextInt/decode(_:). The latter was wasted on systems integers because they call DataInt/normalized() in BinaryInteger.exactly(_:). The situation was similar in InfiniInt, but with normalization after allocating storage (silly). All in all, this should be an improvement in all cases.