[InfiniIntKit][Tests] Use the same numeric values for each T in InfiniInt<T>.
The previous commit makes it easy to create instances from differently sized elements. This lets us keep the numeric value constant while we swap the types in our dynamic loops. I believe this is preferred because it results in different buffer sizes for each type, when there's little reason to believe that upscaled value would behave differently. This is more or less how the Karatsuba multiplication bug was found on 2024-05-15.