Removed: DataInt.Body/load(repeating:).
This method felt out-of-place relative to the other DataInt.Body methods. It is twice as strange when you realize that DataInt is the perfect candidate for this method because it does not need the extra argument. It just happens that implementing this method on DataInt.Body is convenient insofar as it makes this method more accessible. I believe, however, that it would be better to always use DataInt. In general, DataInt.Body methods should not try to extend it. So let's just hope that the compiler sees DataInt(body, repeating: appendix).load() as equivalent—because it is.