[TextInt] Added a T.size <= UX.size encoding fast path.
This fast path skips the DataInt<U8> upsizing step by calling load(as: UX.self) instead. It also simplifies the initial normalization step with a single zero comparison that works because we know the length of the data integer body is at most one UX element in this path. This improves the first million UX encoding benchmark by about 20%. I also maked the DataInt<U8> step as usable from inline instead of inlinable as its impact on big integers is not as relevant.

One thing to note is that the 2s complement of IX.min has the bit pattern of IX.min.magnitude but it is still negative. So I had to reinterpret the integer with Magnitude.init(raw:) before calling load(as: UX.self). I inlined a comment about it, and there exists tests to catch it. Only the most negative signed value has this behavior for the same reason that it is the only signed value that overflows when calling negated().