[InfiniIntKit][Tests] Only pointer-bit and 8-bit elements are interesting.
I removed InfiniInt tests with 16-bit, 32-bit and 64-bit elements in favor of 8-bit and pointer-bit. The removed test were not that interesting and made the tests slower than necessary. What matters, really, is whether the element type is smaller, larger, or pointer-bit. 8-bit is the most extreme smaller case and I'll have to wait for a 128-bit systems integer to land in Swift.