[InfiniIntKit][Todos] Negative and infinite large storage multiplication edge case (#6).
I found a multiplication bug similar to the one in (#3). It's for a similar reason, but when the storage is large this time. If both inputs have bodies filled with 0s and an appendix of 1, then the product should be the combined number of 0s followed by a single 1. This is the only case that does not fit in the combined body allocation. I've added some failing tests that I'll fix soon enough.