[InfiniIntKit] Negative and infinite large storage multiplication edge case (#6).
There's exactly one input combination that does not fit in a (lhs.count + rhs.count) allocation and this commit fixes it. I chose the zero-dropping solution because I prefer the normalization mode of the extended format over the normalization mode of the compact fomat. I also tinkered a bit with the storage model's shift methods and removed the (large, small) multiplication path. The latter was not that different from the (large, large) path and I did not want to maintain the algorithm. If I turn the storage into a (large, small) enum then I might consider a (small, small) fast path.