Improved arbitrary integer exponentiation (#53).
The exponent can be clamped to UX because the allocation limit is IX.max bits. Note that the clamped value must preserve the least significant bit of the original exponent in order to correctly toggle ~0. Also note that only 0, 1, and ~0 can be allocated when the exponent is greater than IX.max. This approach replaces arbitrary integer shifts with systems integer shifts.