Better Divider<T> bounds (#58).
I redid the calculation and it turns out that the rounding error limit is inclusive in both cases (which makes sense); ask me anything. The limit is only exclusive if you assume that there exists a dividend equal to `2^T.size`. It is inclusive since the maximum dividend is `2^T.size-1`.